Graph proof of concept. Missing elements listed below.
Needs standardized dates across APIS. Needs a set of values at the max date
so we can have labels on the right hand side. Might need to consider
switching to a day over day percent difference method, but that should
be done on the API side, not the visualization side.